Исправили типы данных

+++ src/Caclculator.java
@@ -1,5 +1,5 @@
  class Caclculator {
-    public Double calculate(Double first, Integer second, String operation) {
+    public Double calculate(Double first, Double second, String operation) {
         switch (operation) {
             case "+":
                 return first + second;
@@ -9,8 +9,7 @@
                 return first / second;
             case "*":
                 return first * second;
-            default:
-                System.out.println("Такая операция не поддерживается.");
+            default: System.out.println("Такая операция не поддерживается.");
                 return Double.valueOf("-1");
         }
     }
